src/frontend/requirements.txt
python-fasthtml
monsterui
httpx
python-dotenv
uvicorn
pytest
pytest-asyncio
